+++ client/jackson/build.gradle
@@ -4,6 +4,7 @@ apply plugin: 'net.corda.plugins.publish-utils'
 
 dependencies {
     compile project(':core')
+    compile project(':finance')
     compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
     testCompile "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
