+++ core/src/test/kotlin/net/corda/core/contracts/StructuresTests.kt
@@ -6,10 +6,12 @@ import com.nhaarman.mockito_kotlin.whenever
 import org.junit.Test
 import java.io.ByteArrayOutputStream
 import java.io.IOException
+import java.util.*
 import java.util.jar.JarFile.MANIFEST_NAME
 import java.util.zip.ZipEntry
 import java.util.zip.ZipOutputStream
 import kotlin.test.assertEquals
+import kotlin.test.assertNotEquals
 import kotlin.test.fail
 
 class AttachmentTest {
@@ -36,6 +38,38 @@ class AttachmentTest {
             assertEquals("line too long", e.message)
         }
         assertEquals(1, closeCalls)
+    }
+
+}
+
+class UniqueIdentifierTests {
+
+    @Test
+    fun `unique identifier comparison`() {
+        val ids = listOf(UniqueIdentifier.fromString("e363f00e-4759-494d-a7ca-0dc966a92494"),
+                UniqueIdentifier.fromString("10ed0cc3-7bdf-4000-b610-595e36667d7d"),
+                UniqueIdentifier("Test", UUID.fromString("10ed0cc3-7bdf-4000-b610-595e36667d7d"))
+        )
+        assertEquals(-1, ids[0].compareTo(ids[1]))
+        assertEquals(1, ids[1].compareTo(ids[0]))
+        assertEquals(0, ids[0].compareTo(ids[0]))
+        // External ID is not taken into account
+        assertEquals(0, ids[1].compareTo(ids[2]))
+    }
+
+    @Test
+    fun `unique identifier equality`() {
+        val ids = listOf(UniqueIdentifier.fromString("e363f00e-4759-494d-a7ca-0dc966a92494"),
+                UniqueIdentifier.fromString("10ed0cc3-7bdf-4000-b610-595e36667d7d"),
+                UniqueIdentifier("Test", UUID.fromString("10ed0cc3-7bdf-4000-b610-595e36667d7d"))
+        )
+        assertEquals(ids[0], ids[0])
+        assertNotEquals(ids[0], ids[1])
+        assertEquals(ids[0].hashCode(), ids[0].hashCode())
+        assertNotEquals(ids[0].hashCode(), ids[1].hashCode())
+        // External ID is not taken into account
+        assertEquals(ids[1], ids[2])
+        assertEquals(ids[1].hashCode(), ids[2].hashCode())
     }
 
 }
