+++ experimental/src/main/kotlin/net/corda/contracts/universal/Literal.kt
@@ -1,7 +1,7 @@
 package net.corda.contracts.universal
 
-import net.corda.core.contracts.BusinessCalendar
-import net.corda.core.contracts.Frequency
+import net.corda.contracts.BusinessCalendar
+import net.corda.contracts.Frequency
 import net.corda.core.identity.Party
 import java.math.BigDecimal
 import java.time.LocalDate
