+++ experimental/src/test/kotlin/net/corda/contracts/universal/RollOutTests.kt
@@ -1,6 +1,6 @@
 package net.corda.contracts.universal
 
-import net.corda.core.contracts.Frequency
+import net.corda.contracts.Frequency
 import net.corda.core.utilities.DUMMY_NOTARY
 import net.corda.testing.transaction
 import org.junit.Test
