+++ experimental/src/test/kotlin/net/corda/contracts/universal/Swaption.kt
@@ -1,7 +1,7 @@
 package net.corda.contracts.universal
 
-import net.corda.core.contracts.Frequency
-import net.corda.core.contracts.Tenor
+import net.corda.contracts.Frequency
+import net.corda.contracts.Tenor
 import net.corda.core.utilities.DUMMY_NOTARY
 import net.corda.testing.transaction
 import org.junit.Ignore
