+++ finance/src/main/kotlin/net/corda/contracts/clause/Net.kt
@@ -1,6 +1,8 @@
 package net.corda.contracts.clause
 
 import com.google.common.annotations.VisibleForTesting
+import net.corda.contracts.NetCommand
+import net.corda.contracts.NetType
 import net.corda.contracts.asset.Obligation
 import net.corda.contracts.asset.extractAmountsDue
 import net.corda.contracts.asset.sumAmountsDue
