+++ finance/src/main/kotlin/net/corda/contracts/testing/DummyDealContract.kt
@@ -1,12 +1,12 @@
 package net.corda.contracts.testing
 
+import net.corda.contracts.DealState
 import net.corda.core.contracts.Contract
-import net.corda.core.contracts.DealState
 import net.corda.core.contracts.TransactionForContract
 import net.corda.core.contracts.UniqueIdentifier
-import net.corda.core.crypto.*
+import net.corda.core.crypto.SecureHash
+import net.corda.core.crypto.containsAny
 import net.corda.core.identity.AbstractParty
-import net.corda.core.identity.AnonymousParty
 import net.corda.core.identity.Party
 import net.corda.core.transactions.TransactionBuilder
 import java.security.PublicKey
