+++ finance/src/main/resources/net/corda/contracts/LondonHolidayCalendar.txt
