+++ finance/src/main/resources/net/corda/contracts/NewYorkHolidayCalendar.txt
