+++ node/src/test/java/net/corda/node/services/vault/VaultQueryJavaTests.java
@@ -1,37 +1,50 @@
 package net.corda.node.services.vault;
 
-import com.google.common.collect.*;
-import kotlin.*;
-import net.corda.contracts.asset.*;
+import com.google.common.collect.ImmutableSet;
+import kotlin.Pair;
+import net.corda.contracts.DealState;
+import net.corda.contracts.asset.Cash;
 import net.corda.core.contracts.*;
-import net.corda.core.crypto.*;
-import net.corda.core.node.services.*;
-import net.corda.core.node.services.vault.*;
-import net.corda.core.node.services.vault.QueryCriteria.*;
-import net.corda.core.serialization.*;
-import net.corda.core.transactions.*;
-import net.corda.node.services.vault.schemas.*;
-import net.corda.testing.node.*;
+import net.corda.core.crypto.SecureHash;
+import net.corda.core.node.services.Vault;
+import net.corda.core.node.services.VaultService;
+import net.corda.core.node.services.vault.PageSpecification;
+import net.corda.core.node.services.vault.QueryCriteria;
+import net.corda.core.node.services.vault.QueryCriteria.LinearStateQueryCriteria;
+import net.corda.core.node.services.vault.QueryCriteria.VaultQueryCriteria;
+import net.corda.core.node.services.vault.Sort;
+import net.corda.core.serialization.OpaqueBytes;
+import net.corda.core.transactions.SignedTransaction;
+import net.corda.core.transactions.WireTransaction;
+import net.corda.node.services.vault.schemas.VaultLinearStateEntity;
+import net.corda.testing.node.MockServices;
 import org.bouncycastle.asn1.x500.X500Name;
-import org.jetbrains.annotations.*;
-import org.jetbrains.exposed.sql.*;
-import org.junit.*;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.exposed.sql.Database;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Ignore;
+import org.junit.Test;
 import rx.Observable;
 
-import java.io.*;
+import java.io.Closeable;
+import java.io.IOException;
 import java.util.*;
-import java.util.stream.*;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
-import static net.corda.contracts.asset.CashKt.*;
+import static net.corda.contracts.asset.CashKt.getDUMMY_CASH_ISSUER;
+import static net.corda.contracts.asset.CashKt.getDUMMY_CASH_ISSUER_KEY;
 import static net.corda.contracts.testing.VaultFiller.*;
-import static net.corda.core.node.services.vault.QueryCriteriaKt.*;
-import static net.corda.core.node.services.vault.QueryCriteriaUtilsKt.*;
-import static net.corda.core.utilities.TestConstants.*;
-import static net.corda.node.utilities.DatabaseSupportKt.*;
+import static net.corda.core.node.services.vault.QueryCriteriaKt.and;
+import static net.corda.core.node.services.vault.QueryCriteriaUtilsKt.getMAX_PAGE_SIZE;
+import static net.corda.core.utilities.TestConstants.getDUMMY_NOTARY;
+import static net.corda.node.utilities.DatabaseSupportKt.configureDatabase;
 import static net.corda.node.utilities.DatabaseSupportKt.transaction;
-import static net.corda.testing.CoreTestUtils.*;
-import static net.corda.testing.node.MockServicesKt.*;
-import static org.assertj.core.api.Assertions.*;
+import static net.corda.testing.CoreTestUtils.getMEGA_CORP;
+import static net.corda.testing.node.MockServicesKt.makeTestDataSourceProperties;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @Ignore
 public class VaultQueryJavaTests {
