+++ node/src/test/kotlin/net/corda/node/services/vault/VaultQueryTests.kt
@@ -1,15 +1,16 @@
 package net.corda.node.services.vault
 
 import net.corda.contracts.CommercialPaper
+import net.corda.contracts.DealState
 import net.corda.contracts.asset.Cash
 import net.corda.contracts.asset.DUMMY_CASH_ISSUER
 import net.corda.contracts.testing.fillWithSomeTestCash
 import net.corda.contracts.testing.fillWithSomeTestDeals
 import net.corda.contracts.testing.fillWithSomeTestLinearStates
 import net.corda.core.contracts.*
-import net.corda.core.identity.Party
 import net.corda.core.crypto.entropyToKeyPair
 import net.corda.core.days
+import net.corda.core.identity.Party
 import net.corda.core.node.services.Vault
 import net.corda.core.node.services.VaultService
 import net.corda.core.node.services.linearHeadsOfType
