+++ samples/irs-demo/src/main/kotlin/net/corda/irs/contract/IRS.kt
@@ -1,5 +1,6 @@
 package net.corda.irs.contract
 
+import net.corda.contracts.*
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
