+++ samples/irs-demo/src/main/kotlin/net/corda/irs/contract/IRSUtils.kt
@@ -1,9 +1,9 @@
 package net.corda.irs.contract
 
+import net.corda.contracts.Tenor
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import net.corda.core.contracts.Amount
-import net.corda.core.contracts.Tenor
 import net.corda.core.serialization.CordaSerializable
 import java.math.BigDecimal
 import java.util.*
